updates for esmda

+++ run_fg_ens.sh
@@ -6,8 +6,10 @@ echo "deleting existing files..."
 nanal=1
 while [ $nanal -le $nanals ]; do
     charnanal="mem`printf %03i $nanal`"
-    /bin/rm -f ${datapath}/${analdatep1}/sfg_${analdatep1}*${charnanal}
-    /bin/rm -f ${datapath}/${analdatep1}/bfg_${analdatep1}*${charnanal} 
+    if [ $nliteration -eq $nliterations ]; then
+       /bin/rm -f ${datapath}/${analdatep1}/sfg_${analdatep1}*${charnanal}
+       /bin/rm -f ${datapath}/${analdatep1}/bfg_${analdatep1}*${charnanal} 
+    fi
     nanal=$((nanal+1))
 done
 fi
@@ -18,10 +20,10 @@ alldone='no'
 echo "${analdate} compute first guesses `date`"
 while [ $alldone == 'no' ] && [ $niter -le $nitermax ]; do
     if [ $niter -eq 1 ]; then
-    ${enkfscripts}/${fg_gfs} > ${current_logdir}/run_fg.iter${niter}.out 2>&1
+    ${enkfscripts}/${fg_gfs} > ${current_logdir}/run_fg${nliteration}.iter${niter}.out 2>&1
     exitstat=$?
     else
-    ${enkfscripts}/${fg_gfs} > ${current_logdir}/run_fg.iter${niter}.out 2>&1
+    ${enkfscripts}/${fg_gfs} > ${current_logdir}/run_fg${nliteration}.iter${niter}.out 2>&1
     exitstat=$?
     fi
     if [ $exitstat -eq 0 ]; then
